implementation of getAverage method

+++ src/main/java/com/telerikacademy/projectattempt/data/CarRepository.java
@@ -9,5 +9,5 @@ public interface CarRepository {
 
     public List<Car> listAll();
 
-    public int getAverage(Car car);
+    public Double getAverage(Car car);
 }
